pyproject.toml
[project]
name = "uv-test-20260226"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "google-cloud-aiplatform>=1.139.0",
    "google-cloud-storage>=3.9.0",
    "google-genai>=1.65.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.4",
]

[tool.ruff]
# 1行の最大文字数
# デフォルトの88だとインデントがスペース4個分なので折り返されてしまう
line-length = 100

# チェック対象とするPythonのバージョン
target-version = "py312"

[tool.ruff.lint]
# 有効にするルールのカテゴリを指定
select = [
    "E",  # pycodestyle エラー (基本ルール)
    "F",  # Pyflakes (未使用インポートなど)
    "I",  # isort (インポート順の整理)
    "B",  # flake8-bugbear (バグになりやすい書き方の検知)
    "UP", # pyupgrade (古いPython構文を新しいものに自動変換)
    "N",  # pep8-naming (命名規則)
]

# あえて無視するルール（必要に応じて追加）
ignore = []

[tool.ruff.lint.isort]
# インポート順を整理する際の詳細設定
# たとえば特定のライブラリを常に一番上にしたい場合などに使います
known-first-party = ["your_project_name"]

[tool.ruff.format]
# フォーマット設定
quote-style = "double"        # ダブルクォートに統一
indent-style = "space"        # スペースでインデント
skip-magic-trailing-comma = false

[tool.mypy]
# 型ヒントがない関数もチェック対象にする（少し厳しめ）
check_untyped_defs = true

# インポート先のライブラリに型定義がなくてもエラーにしない（SDK利用時に重要）
ignore_missing_imports = true

# 戻り値に Any 型が含まれる場合に警告（任意。厳しくしたい場合は true）
warn_no_return = true

# 実行時のベースとなるPythonバージョン
python_version = "3.12"

# 無視するディレクトリ（仮想環境などは除外）
exclude = [
    ".venv",
    "tests",
]

# srcレイアウト配下のモジュール重複読み込みエラーを防止
# 例えば、src/core/ai.py が "ai" と "src.core.ai" の両方で認識されるエラーを防ぐ
explicit_package_bases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]
